* Boot into a fully functional stage 2. * Stage 2 init: handle non-read-only roots.
svn path=/nixu/trunk/; revision=7014

+++ test/boot-stage-2-init.sh
@@ -2,10 +2,12 @@
 
 # !!! copied from stage 1; remove duplication
 
+
 # Print a greeting.
 echo
 echo "<<< NixOS Stage 2 >>>"
 echo
+
 
 # Set the PATH.
 export PATH=/empty
@@ -16,28 +18,46 @@ for i in @path@; do
     fi
 done
 
-# Mount special file systems.
-mount -t tmpfs none /etc -n # to shut up mount
-touch /etc/fstab # idem
+
+# Mount special file systems, initialise required directories.
+
+if test -z "@readOnlyRoot@"; then
+    #rootDev=$(grep "/dev/.* / " /proc/mounts | sed 's/^\([^ ]*\) .*/\1/')
+    mount -o remount,rw /dontcare / # !!! check for failure
+fi
+
+needWritableDir() {
+    if test -n "@readOnlyRoot@"; then
+        mount -t tmpfs none $1 $3
+        chmod $2 $1
+    else
+        mkdir -m $2 -p $1
+    fi
+}
+
+needWritableDir /etc 0755 -n # to shut up mount
+
+test -e /etc/fstab || touch /etc/fstab # idem
+
 mount -t proc none /proc
 mount -t sysfs none /sys
-mount -t tmpfs none /dev
-mount -t tmpfs none /tmp
-mount -t tmpfs none /var
-mount -t tmpfs none /nix/var
+needWritableDir /dev 0755
+needWritableDir /tmp 01777
+needWritableDir /var 0755
+needWritableDir /nix/var 0755
 
-mkdir -p /nix/var/nix/db
-mkdir -p /nix/var/nix/gcroots
-mkdir -p /nix/var/nix/temproots
+mkdir -m 0755 -p /nix/var/nix/db
+mkdir -m 0755 -p /nix/var/nix/gcroots
+mkdir -m 0755 -p /nix/var/nix/temproots
+
 
 # Ensure that the module tools can find the kernel modules.
 export MODULE_DIR=@kernel@/lib/modules/
 
-# Create device nodes in /dev.
-#source @makeDevices@
 
 # Start udev.
 udevd --daemon
+
 
 # Let udev create device nodes for all modules that have already been
 # loaded into the kernel (or for which support is built into the
@@ -45,19 +65,22 @@ udevd --daemon
 udevtrigger
 udevsettle # wait for udev to finish
 
+
 # Start syslogd.
-mkdir -p /var/run
+mkdir -m 0755 -p /var/run
 #mkdir -p /var/log
 #touch /var/log/messages
 echo "*.* /dev/tty10" > /etc/syslog.conf
 echo "syslog 514/udp" > /etc/services # required, even if we don't use it
 @sysklogd@/sbin/syslogd &
 
+
 # Try to load modules for all PCI devices.
 for i in /sys/bus/pci/devices/*/modalias; do
     echo "Trying to load a module for $(basename $(dirname $i))..."
     modprobe $(cat $i)
 done
+
 
 # Bring up the network devices.
 modprobe af_packet
@@ -71,17 +94,25 @@ for i in $(cd /sys/class/net && ls -d *); do
     fi
 done
 
+
 # login/su absolutely need this.
-touch /etc/login.defs 
+test -e /etc/login.defs || touch /etc/login.defs 
+
 
 # Enable a password-less root login.
-echo "root::0:0:root:/:@shell@" > /etc/passwd
-echo "root:*:0" > /etc/group
+if ! test -e /etc/passwd; then
+    echo "root::0:0:root:/:@shell@" > /etc/passwd
+fi
+if ! test -e /etc/group; then
+    echo "root:*:0" > /etc/group
+fi
+
 
 # Set up inittab.
 for i in $(seq 1 6); do 
     echo "$i:2345:respawn:@mingetty@/sbin/mingetty --noclear tty$i" >> /etc/inittab
 done
+
 
 # Show a nice greeting on each terminal.
 cat > /etc/issue <<EOF
@@ -92,6 +123,7 @@ You can log in as \`root'.
 
 
 EOF
+
 
 # Additional path for the interactive shell.
 for i in @extraPath@; do
@@ -106,11 +138,14 @@ export PATH=$PATH
 export MODULE_DIR=$MODULE_DIR
 EOF
 
+
 # Set the host name.
 hostname nixos
 
+
 # Start an interactive shell.
 #exec @shell@
+
 
 # Start init.
 exec init 2
